+++ components/serverless-function/serverless.js
@@ -3,11 +3,9 @@ const execa = require("execa");
 const { Component } = require("@serverless/core");
 const { trackComponent } = require("@webiny/tracking");
 
-const component = "@webiny/serverless-function";
-
 class ServerlessFunction extends Component {
     async default({ track, ...inputs } = {}) {
-        await trackComponent({ track, context: this.context, component });
+        await trackComponent({ track, context: this.context, component: __dirname });
 
         const functionRoot = path.join(this.context.instance.root, inputs.root);
 
@@ -25,15 +23,27 @@ class ServerlessFunction extends Component {
         }
 
         const lambda = await this.load("@serverless/function");
-        return await lambda(inputs);
+        const output = await lambda(inputs);
+
+        this.state.output = output;
+        await this.save();
+
+        return output;
     }
 
     async remove({ track, ...inputs } = {}) {
-        await trackComponent({ track, context: this.context, component, method: "remove" });
+        await trackComponent({
+            track,
+            context: this.context,
+            component: __dirname,
+            method: "remove"
+        });
 
         const lambda = await this.load("@serverless/function");
+        await lambda.remove(inputs);
 
-        return await lambda.remove(inputs);
+        this.state = {};
+        await this.save();
     }
 }