+++ components/serverless-apollo-gateway/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.5](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-apollo-gateway@0.1.4...@webiny/serverless-apollo-gateway@0.1.5) (2019-10-08)
+
+
+### Bug Fixes
+
+* add component version to tracking. ([6fdaad9](https://github.com/Webiny/webiny-js/commit/6fdaad9))
+* empty component state on component removal. ([2dc7899](https://github.com/Webiny/webiny-js/commit/2dc7899))
+
+
+
+
+
 ## [0.1.4](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-apollo-gateway@0.1.3...@webiny/serverless-apollo-gateway@0.1.4) (2019-10-07)
 
 **Note:** Version bump only for package @webiny/serverless-apollo-gateway